finalise tournament controller
