+++ purge_tournament/assets/styles/sass/style.scss
@@ -1,3 +1,4 @@
 @import "reset";
 @import "vars";
+@import "partials/_header";
 @import "tournaments/create-tournament";
