+++ purge_tournament/templates/public/partials/_header.phtml
@@ -1,6 +1,7 @@
 <header>
     <nav>
         <ul>
+            <li><a href="/res03-projet-final/purge_tournament/"><figure><img src="./assets/img/purgelogo.png" alt="Logo représentant un nuage atomique avec le nom de l'entreprise"></img></figure></a></li>
             <li><a href="/res03-projet-final/purge_tournament/about-us">About Us</a></li>
             <li><a href="/res03-projet-final/purge_tournament/schedule">Schedule</a></li>
             <li><a href="/res03-projet-final/purge_tournament/contact">Contact</a></li>
